- setup & verify phrase in different components
